A lot of work to clean up the pick email screen.
Now making the "add email" portion show in its own screen so that things look cleaner.

Adding room for 5 email addresses.

Putting all the buttons on the same line to maximize room for email addresses.

Changing the padding/margins of the email addresses so that the glow around the buttons is not cut off.