Trying to streamline the sync process.
When an identity's key or cert is invalidated, we no longer remove it from storage but instead remove its public key, private, and cert.  Adding invalidateEmail to storage, but may move this to identities as it seems like too much info for storage.  Updating the controllers to call syncEmails instead of syncEmailKeypairs.