Fix the pick_email test failure when all tests are run.
The problem stemmed from the DOM was not completely reset between every test.  When the manage page inserted email addresses into its portion of the DOM, it added email addresses with elements with the same ID as the pick email controller.  Instead of ensuring we clean wipe these elements between every test, rely on QUnit's ability to reset a portion of the DOM to its initial state.

* Add the `qunit-fixture` class to the bit of DOM that should be reset between every test run.
* Add the testHelpers.setup/testHelpers.teardown functions to several suites that manually called emtpy() on DOM elements.

issue #1156