Add note-to-ourselves re: onready's deprecation
onready is in a weird state of limbo: we know we're going to deprecate it, we
know what we're going to replace it with, and we know we want to discourage its
use for the moment, but we haven't actually planned, implemented, or announced
its replacement.

We haven't announced its deprecation, either.

Adding two small comments in-code to ensure that we don't inadvertantly remove
this without giving RPs proper notice. After all, if onready stops firing, we
seriously break sites that rely on it.

See GH-2017 for more info.