Bug 793579: check issuer for primary-address operations.
When "primary" addresses (which come from browserid-aware IdPs) are
associated with a Persona account, users can perform certain
operations (including add_email) with a primary-backed assertion. This
changes primary.verifyAssertion(), which checks such assertions, to make
sure their .issuer field is actually allowed to speak for the address in
question.

It also tightens up the way .issuer is handled in the verifier. In
particular, the verifier will no longer accept a secondary-backed
assertion for domains that run a primary IdP. The verifier will perform
an HTTPS .well-known lookup on each verification request to detect IdP
support.

The .issuer handling is still a bit looser than the spec probably calls
for. If domain A has a .well-known which delegates to B (e.g.
{authority: B}), then certificates with principal.email=foo@A are
supposed to have .issuer=B, but this code will still accept certificates
with .issuer=A.

proxy-idp-test.js fails, I think because it is using the local secondary
to create a "primary" assertion, then exercising auth_with_assertion
with that cert, and the rules are now tighter. To fix this, I think
something must be done with g_shim_cache.