Bug 793579: check issuer for primary-address operations.
When "primary" addresses (which come from browserid-aware IdPs) are
associated with a Persona account, users can perform certain
operations (including add_email) with a primary-backed assertion. This
changes primary.verifyAssertion(), which checks such assertions, to make
sure their .issuer field is actually allowed to speak for the address in
question.

proxy-idp-test.js fails, I think because it is using the local secondary
to create a "primary" assertion, then exercising auth_with_assertion
with that cert, and the rules are now tighter. To fix this, I think
something must be done with g_shim_cache.