primary.ps: pass principalDomain through to everything
We now pass two arguments into each getWellKnown() and
parseWellKnownBody() call. The "fromDomain" argument changes on each
delegated lookup (it indicate which server we're fetching
.well-known/browserid from *right now*), while "principalDomain" remains
constant (and indicates which domain we really care about, regardless of
who we're currently asking about them).

This shouldn't change any behavior. It just paves the way for a later
change to add principalDomain as a query arg.