Fix getSilentAssertion for B2G.
Firefox OS 1.0.1 keeps one copy of the communication iframe around. When it loads, it fetches a copy of session context and then caches it. Whenever the user signs in to a tab, the communication iframe's session context cache is not updated. To avoid updates to B2G code, force a refresh of session context whenever calling user.getSilentAssertion.