Make sure all expected KPIs are sent.
~ 4% of the KPI blobs were missing:
* number_emails
* number_sites_signed_in
* number_sites_remembered
* orphaned

The problem is that occasionally, the window would unload before the onbeforeunload handler could set the last KPIs.

Since KPIs are now sent in the communication iframe instead of in the dialog, the last remaining KPIs are set there before sending the blobs.

orphaned now means, 100% of the time, no assertion was generated in the dialog.

fixes #2280