Allow new users to edit the email field after they cancel from set password.
The assumption that just because email was passed to the controller that the email was immutable was invalid. When a user cancels from the set password state and the authentication controller is restarted, the previous email is passed in to the controller and is editable.

Instead of trying to be too fancy, pass in whether the email is mutable or not.

fixes #3689