Explicitly set rejectUnauthorized to false when using https.request()
This is the default in node 0.10 but not in 0.8. That older version
needs both `agent: false` and `rejectUnauthorized: true` before it
will reject invalid TLS certificates.

http://nodejs.org/api/https.html#https_https_request_options_callback

In production, this is normally not needed because our squid proxy
is also set to reject invalid certificates, but setting it like
this means less surprises when we upgrade to node 0.10 and protects
anybody that might be running our code without a proxy on the old
version of node.

https://bugzilla.mozilla.org/show_bug.cgi?id=901650