Fix certs being requested from the fallback IdP for addresses supported by primary IdPs.
`type` has not been stored on email address information that is stored in localStorage for many months. User.getAssertion was not updated. Instead of looking at the address info stored in localStorage, User.addressInfo must be called instead, and the information fetched from the backend.

* Update the getAssertion tests to remove the `type` field when calling storage.addEmail
* Update user.getAssertion to get an email type from the information returned from User.addressInfo

Partial fix for Bugzilla #912001
fixes #3625