Simplify the way the statsd middleware works.
No need to call `statsd.setupMiddleware`, just call `app.use(statsd());`

If statsd is configured, statsd() will return a middleware to send responses
to statsd. If statsd is not configured, an empty middleware that calls `next()`
is returned. This allows users of the middleware to not worry about whether statsd
is configured before adding the middleware.